requirements.txt
numpy==1.20
scikit-learn==0.24.2
matplotlib==3.4.3
seaborn==0.11.2
umap-learn==0.5.1
numba==0.54.0
